backend/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
django-filter
pillow